+++ src/_compiled_header.js
@@ -1,0 +1,8 @@
+// Local Toplevel
+// ==============
+// pfraze 2013
+
+if (typeof this.local == 'undefined')
+	this.local = {};
+
+(function() {