+++ src/aliases.js
@@ -1,0 +1,4 @@
+// Aliases to common functions
+local.dispatch = local.web.dispatch;
+local.subscribe = local.web.subscribe;
+local.navigator = local.web.navigator;